requirements.txt
streamlit==1.50.0
httpx>=0.25.0
python-dotenv>=1.0.0
langchain-core==1.0.0